Tighten shared plugin install boundaries
